session: avoid KnownHostsCommand when unsupported
We want to run our tests on centos-8-stream, but the version of SSH
available there doesn't known about KnownHostsCommand.

Use the -G option to detect the available set of options that ssh(1)
supports.  This is a lot more flexible than hard-coding versions.  If
KnownHostsCommand isn't supported, we simply ignore the handle_host_key
parameter.

Move StrictHostKeyChecking=yes to only be enabled if we are using our
KnownHostsCommand.  Now we're going to get a normal "this key is not
known, do you trust it..." type prompt as an askpass request.

Add a detection for that in our testcase and redirect to the hostkey
handler. This works in this case because our testcase doesn't care about
the actual value of the key (which isn't available via this route).