session: catch ProcessLookupError on .kill()
0017f9b603 introduced a .kill() call on the ssh process if
authentication failed, in order to avoid a deadlock.

Even though the subprocess code goes out of its way not to emit a
ProcessLookupError, the asyncio code will explicitly raise it if it
knows that the child process has already exited.

This shows up from time to time in our unit tests since the last commit.

In any case, let's just catch and ignore the exception.