Introduce better detection of errors
Add some "magic detection" code to make sense of the messages that ssh
writes to its stderr as it exits.  We support mapping OSError-type
errors for network failures, gaierror-type errors for DNS problems, as
well as a couple of ssh-specific errors for hostkey or
authentication-related problems.

Tweak the tests accordingly.