session: improve feature detection
The idea to use `ssh -G` to get a list of supported features only works
for features which have non-`NULL` default values.  If a particular
configuration option is `NULL` (ie: `none`) then it is skipped in the
output.

This happens to be the case for KnownHostsCommand, which means that our
attempt to detect it was always coming back false, hitting the fallback
path and the workaround in the tests, but failing in real-life use.

A quick read of the code of OpenSSH seems to suggest that there's no
other way to get a list of valid options, so change our approach: we can
query a specific option by attempting to set it and see if we get an
error or not.

Put this under tests so we know that it's working properly.  This will
also detect possible future changes to the behaviour of ssh.