test: Check large environment
With a bunch of `$XDG_DATA_DIRS`, long `$PATH` and `$LS_COLORS` it is
easy to exceed 4KB in the environment. Before commit 85c805c61a13, this
made the agent (and most of the tests) crash with

    SyntaxError: unterminated string literal (detected at line 1)

when reading the request.

Add a unit test with a large environment variable to ensure that this
does not regress again.