interaction_client: Support cancelling through EOF on stdin
This is how cockpit's channels work -- closing them (when the user
cancels an authentication dialog, for example) closes the spawned
program's stdin and SIGTERMs it. Our interaction agent never sees
SIGTERM, as that goes to the program that spawned it (like sudo or ssh),
and as sudo ignores SIGTERM, the stdin EOF is our only chance to get
signaled about "I want to cancel".

However, we need to be careful to not immediately exit if stdin is
immediately EOF, as it is the case for /dev/null.