Improve the ferny interaction protocol
Make the interaction protocol clearer and more robust, and document it.

This is an incompatible protocol change: instead of writing the command
as in-band data in the stderr socket, we send a pipe as an extra fd and
write the command over that.  As in-band data we send a single '\0'.

The main reason for this change was the existence of a bug in the
current approach: in general, it's not possible to reliably know which
in-band data was sent along with which control message, except in the
case that only a single byte is sent.

Sending only a single byte of in-band data also avoids another somewhat
more pedestrian issue: in case the socket buffer was full, and we
received a signal during the sendmsg() call in `ferny-askpass`, we'd
have to handle the possibility of a partial retransmission.  If we send
a single byte, then the write either succeeds or it doesn't — and if it
doesn't, then `SA_RESTART` (or failing that, Python) will retry the
syscall.

At the same time as making this change, take the chance to add a clearer
conceptual boundary between the idea of sending "commands" and all of
the extra handling that we need to do to implement askpass.  Also,
remove the conceptual boundary between "local" and "remote" commands.

`InteractionAgent` is now more or less responsible for implementing the
low-level "sending commands" layer, and `InteractionResponder` is
responsible for the implementation of the commands (including askpass).
The `Interaction` helper class no longer exists — most of what it
contained and did is now contained entirely in the _askpass_command()
function.