agent: fix our wait_readable() for cancellation
If the task running wait_readable() is cancelled in the same loop
iteration as the fd polls readable (which could happen when a process
exits, for example) then the future will get cancelled, and then
set_result() will be called on it, leading to an InvalidStateError.

Solve this by adding a small wrapper function to check first for
cancellation on the future.