test: add a couple of test cases for agent cancellation
One of these cases is currently failing.  That's because if we close the
agent before we get an askpass interaction, the fds from the interaction
will stay queued on the socketpair forever, and will never get closed.
We can fix that by closing our end, but we should probably figure out a
better way.  Mark it xfail so we remember to fix it.