agent: close our socket after communicate()
Enclose our entire communicate() function in a with: block to ensure
that our socket gets properly closed at the end of the communication.

This has the side-effect of making sure that anyone else who tries to
write to stderr will get EPIPE.  That means that a new incoming askpass
request won't get suck if communicate() has been cancelled, which fixes
an xfail.

This is not exactly what I wanted to do here — in theory we could
continue to collect stderr messages.  But we don't do that right now
anyway, so this ought not to hurt existing users.  If we want to do that
in the future, we can consider a different solution.