test: Check session with handle_host_key=False
Make `handle_host_key` a parameter of run_test(), and check the three cases
(unknown, known, and changed host key) with `handle_host_key=False`. We don't
expect any do_hostkey() interactions in any of the three cases; but it
unexpectedly happens if the key is not known, so xfail that test and add
a comment to the mock responder why/where it happens.