test: Check do_askpass() and do_hostkey() arguments
These are highly relevant for the UI, let's make sure that they work as
intended. This uncovered an issue: With a changed hostkey there is
currently no do_hostkey() prompt for accepting the new key. Cockpit
expects that, so we will eventually have to fix this. Mark it with
"FIXME" for now.

These replace the `print()` debugging statement.

Also move the MockResponder constructor to the top, to make it more
obvious that this class has one.