test: Fix test_known_host_good
I misunderstood `KnownHostsCommand` in commit c4fdd8: It will be called
with a `HOSTNAME` command *even if* the known_hosts files already have a
correct key, apparently so that it can add additional host keys. Thus
do_hostkey() needs to be prepared for that.

This means that we can never use `handle_host_key=True` in the initial
connection with the `KnownHostsCommmand` approach, as it's impossible to
tell apart the "unknown/known/changed" situations in do_hostkey(). But
we are planning to get rid of KnownHostsCommmand with some API redesign
anyway, so let's just fix the test for now: Instead of assuming that
do_hostkey() does not get called, let it reject all keys, so that the
connection is forced to take the known_host file into account. This
works fine.