session: Stop disabling known host files for handle_host_key
Doing that prevents SSH from detecting whether the given host is
unknown, known, or changed the host key. But we want that to send
correct exceptions -- it is really weird to consider every host as
"unknown" in `handle_host_key` mode.

The unit tests paper over this by explicitly setting UserKnownHostsFile
to a temporary one on the command line. This is unfortunately difficult
to do in a better way, as SSH uses `getpwent()` to compute ~/.ssh
instead of looking at `$HOME`, so the only way to test that properly in
a unit test would be to move the actual ~/.ssh away during the tests --
and that is rather unfriendly and brittle for developer machines.

Note that this does *not* stop SSH from calling the `KnownHostsCommand`
-- it will always do that even for existing known_hosts entries.