agent: add another new way to get the client code
Previously we were exporting the somewhat awkward-to-use
COMMAND_TEMPLATE which beiboot was importing for its own purposes.  This
was always an awkward API and beiboot importing it directly caused
problems with vendoring beiboot and ferny together in a third package
(Cockpit).

Instead of COMMAND_TEMPLATE, provide a new BEIBOOT_GADGETS dictionary to
be used in the upcoming beiboot support for "gadgets" as part of
constructing the first-stage bootloader.