agent: support multiple responder objects
Create a superclass for InteractionResponder called InteractionHandler,
which is the interface consumed by InteractionAgent.  The handler now
defines the exact list of commands that it's capable of handling and
only receives those commands dispatched to it.

The constructor for the InteractionAgent now accepts InteractionHandler
but it's optional, and you can also add handlers later via a new
.add_handler() function.