interaction_agent: Classify SSH askpass prompts
Create a new ssh_askpass.py which contains logic for parsing ssh-askpass
prompts issued by OpenSSH.  We have a new class hierarchy and
double-dispatch mechanism for a half dozen or so different types of
prompts.

Rework the base askpass handling a bit to make room for that work.  The
biggest change to endusers is that ferny.InteractionResponder no longer
exists and needs to be replaced by ferny.AskpassHandler or
ferny.SshAskpassResponder depending on if you want the categorization
API or not.

Add some regexp-based tests to test_regexp.py for testing this new
functionality.