session: compute runtime dir at Session.connect()
We were previously computing the runtime directory at the file
top-level, which means that it was sampling the XDG_RUNTIME_DIR
environment variable at the time the script was parsed.

That worked fine for normal usage but conflicted with our attempts to
use a temporary directory for our tests.  The temporary directory that
we were using for our tests was simply ignored — that's also why it was
empty at the end, and didn't contain the 'ferny' directory as expected.

Let's move the logic into the Session.connection() function so that it
checks XDG_RUNTIME_DIR after the tests have patched it.  Fix the broken
assert to expect that the 'ferny' directory *does* exist at the end, and
add usage of the directory to a couple of extra tests that were missing
it before.

The reason that this always worked up until now is that we've only ever
run tests in environments that already had XDG_RUNTIME_DIR set, and we
ended up using that directory.  Running under 'tox' (where we have a
cleaner environment) revealed the problem.