all: new approach to type hints compatibility
Nothing inside of ferny makes use of runtime inspection of static typing
hints, so let's take advantage of the fact that type hints are allowed
to be quoted to avoid some compatibility issues.

We now try to use the "latest and greatest" version of the type hints
available to us, quoting the entire type declaration where it would
otherwise cause problems with older Python versions.

Add some "banned import" rules to our ruff config to make sure we get
this right.