Add new class FernyTransport
This is a asyncio.Transport which handles most of the "logistics" of
spawning SSH under a running InteractionAgent and dealing with all of
the exceptional situations which may arise from that.  It's now
considered the preferred API of ferny for those looking to connect
asyncio.Protocol objects (such as Cockpit).