interaction_agent: rewrite in callback style
The interaction_agent is currently more or less implemented via the
logic in the async communicate() function.  We do a loop which waits for
the stderr fd to become readable and then handles the data that comes in
to it.

Change that around: let's instead register a callback to be called when
the stderr fd becomes readable and run handlers from there.  It means we
no longer have a top-level task, so we need to create tasks for running
the handlers (which keep their previous async API).  That presents an
advantage, though: since we now run each handler in a separate task, the
logic for cancellation gets a whole lot easier.

So far this looks like a fairly pointless refactor, but it will enable
non-task-based users of this API.  Those will follow in the next commits.