ssh_errors: ignore \r output by SSH
For some strange reasons possibly related to RFC compliance, ssh(1)
outputs its stderr messages with a carriage return added at the end of
each line, just before the newline.

Strip that out before we try to match the error messages, or nothing is
going to match.

Drop a stray debugging print() while we're at it.